requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-genai>=0.8.0
yt-dlp>=2024.1.0
